Address stimulant review cleanup
