+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/MedicalChemistry/AnyStatusMetabolismReagentSelector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 84ec5b1122c34329b6de5679b86b00c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
