+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/MedicalChemistry/ApplyStatusMetabolismReaction.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using Chemistry;
 using UnityEngine;
@@ -9,42 +8,6 @@ using US13.Systems.StatusesAndEffects;
 
 namespace US13.HealthV2.Living.MedicalChemistry
 {
-	public interface IStatusMetabolismReagentSelector
-	{
-		bool HasMatch(ReagentMix reagentMix);
-		void ConsumeMatches(ReagentMix reagentMix, float maxReactQuantity, float metabolismMultiplier);
-	}
-
-	[Serializable]
-	public class AnyStatusMetabolismReagentSelector : IStatusMetabolismReagentSelector
-	{
-		public List<Reagent> Reagents = new();
-
-		public bool HasMatch(ReagentMix reagentMix)
-		{
-			if (reagentMix == null) return false;
-			foreach (var reagent in Reagents)
-			{
-				if (reagent != null && reagentMix[reagent] > 0) return true;
-			}
-			return false;
-		}
-
-		public void ConsumeMatches(ReagentMix reagentMix, float maxReactQuantity, float metabolismMultiplier)
-		{
-			if (reagentMix == null || reagentMix.Total == 0) return;
-			var reactionScale = maxReactQuantity / reagentMix.Total;
-			foreach (var reagent in Reagents)
-			{
-				if (reagent == null) continue;
-				var untouchedMultiple = reagentMix[reagent];
-				if (untouchedMultiple <= 0) continue;
-				var reactionMultiple = Mathf.Min(untouchedMultiple * reactionScale * metabolismMultiplier, untouchedMultiple);
-				reagentMix.Subtract(reagent, reactionMultiple);
-			}
-		}
-	}
-
 	[CreateAssetMenu(fileName = "ApplyStatusMetabolismReaction", menuName = "ScriptableObjects/Chemistry/Reactions/ApplyStatusMetabolismReaction")]
 	public class ApplyStatusMetabolismReaction : MetabolismReaction
 	{
