+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/MedicalChemistry/IStatusMetabolismReagentSelector.cs
@@ -1,0 +1,10 @@
+using Chemistry;
+
+namespace US13.HealthV2.Living.MedicalChemistry
+{
+	public interface IStatusMetabolismReagentSelector
+	{
+		bool HasMatch(ReagentMix reagentMix);
+		void ConsumeMatches(ReagentMix reagentMix, float maxReactQuantity, float metabolismMultiplier);
+	}
+}
