+++ UnityProject/Assets/Scripts/US13/HealthV2/Living/MedicalChemistry/IStatusMetabolismReagentSelector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 74ecf809fe394d09b03c603145a5760e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
