+++ UnityProject/Assets/Tests/StatusAndEffectsFramework/ExpirableStatusEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 63da5233fff24f5c84d460db9f6bcd43
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
