+++ UnityProject/Assets/Tests/StatusAndEffectsFramework/ImmediateStatusEffect.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: beb8e0ab788949e7b42d34e0ab3f2ebe
+guid: c2e7e6401e974ab9aa511ed2f1b5043c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+  userData:
+  assetBundleName:
+  assetBundleVariant:
