+++ UnityProject/Assets/Tests/StatusAndEffectsFramework/MockStatus.cs
@@ -1,0 +1,9 @@
+using US13.Systems.StatusesAndEffects;
+
+namespace Tests.StatusAndEffectsFramework
+{
+	public class MockStatus : StatusEffect
+	{
+
+	}
+}
