+++ UnityProject/Assets/Tests/StatusAndEffectsFramework/MockStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d8b8bcca57f4f38a5748fe0705b2f3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
