+++ UnityProject/Assets/Tests/StatusAndEffectsFramework/StackableStatusEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 31c7c393d6e648d9a6b427e55980459c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
